chore: relocate CLI entry points
Move *_main.py files into scripts/ package and update pyproject, imports, and docs.
