+++ scripts/table_quality_main.py
