+++ tests/test_cli_run_id.py
@@ -6,8 +6,8 @@ from pathlib import Path
 
 import pytest
 
-import csv_utils_main
 from library.logging_setup import LoggerConfig, configure_logger
+from scripts import csv_utils_main
 
 
 def test_cli_run_id(capfd: pytest.CaptureFixture[str], tmp_path: Path) -> None:
