+++ tests/test_mapper_logging.py
@@ -10,9 +10,9 @@ from typing import Any
 
 import pandas as pd
 
-import mapper_main
 from library.cli import LoggerConfig, configure_logger
 from library.config import Config
+from scripts import mapper_main
 
 
 def test_mapper_library_has_no_logging_side_effect() -> None:
