+++ tests/test_print_config.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import yaml
 
-from table_quality_main import main
+from scripts.table_quality_main import main
 
 
 def test_print_config_cli(tmp_path: Path, monkeypatch, capsys) -> None:
